postgres/requirements.txt
psycopg2
python-decouple